requirements.txt
numpy==2.4.3
pandas==3.0.1
pyserial==3.5
scikit-learn==1.8.0
scipy==1.17.1
